Other layers

+++ src/net/ethernaught/packet/layers/inter/Layer.java
@@ -4,7 +4,7 @@ public interface Layer {
 
     byte[] getBytes();
 
-    long length();
+    int length();
 
-    long computeLength();
+    int computeLength();
 }
